// frontend/src/interfaces/individual/IEmployee.ts
export interface EmployeesInterface  {
	ID?:                number;
    FirstName?: 		string;
    LastName?:  		string;    	
	Birthday?: 		    string; 	
	Address?:			string;		
	Tel?:				string;		
    Email?:     		string;    	
    Password?:  		string;   	
    GenderID?:  		number;      	
	JobPositionID?:  	number;      	
}